+++ src/apps/settings/modules/information/infrastructure.tsx
@@ -1,8 +1,9 @@
 import { exit, relaunch } from '@tauri-apps/plugin-process';
-import { Button, Switch } from 'antd';
+import { Button, Select, Switch } from 'antd';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
+import { UpdateChannel } from 'seelen-core';
 
 import { EnvConfig } from '../shared/config/infra';
 import cs from './infra.module.css';
@@ -17,7 +18,7 @@ export function Information() {
   const [isMsixBuild, setIsMsixBuild] = useState(false);
 
   const devTools = useSelector(newSelectors.devTools);
-  const betaChannel = useSelector(newSelectors.betaChannel);
+  const updaterSettings = useSelector(newSelectors.updater);
 
   const dispatch = useDispatch();
   const { t } = useTranslation();
@@ -30,8 +31,8 @@ export function Information() {
     dispatch(RootActions.setDevTools(value));
   }
 
-  function onToggleBetaChannel(value: boolean) {
-    dispatch(RootActions.setBetaChannel(value));
+  function onChangeUpdateChannel(channel: UpdateChannel) {
+    dispatch(RootActions.setUpdater({ ...updaterSettings, channel }));
   }
 
   return (
@@ -68,8 +69,13 @@ export function Information() {
           <Switch value={devTools} onChange={onToggleDevTools} />
         </SettingsOption>
         <SettingsOption>
-          <span>{t('beta_channel.enable')}</span>
-          <Switch value={!isMsixBuild && betaChannel} onChange={onToggleBetaChannel} disabled={isMsixBuild}/>
+          <span>{t('update.channel')}</span>
+          <Select
+            value={updaterSettings.channel}
+            disabled={isMsixBuild}
+            onChange={onChangeUpdateChannel}
+            options={Object.values(UpdateChannel).map((c) => ({ value: c, label: c }))}
+          />
         </SettingsOption>
       </SettingsGroup>