+++ src/apps/settings/modules/shared/store/app/StateBridge.ts
@@ -32,7 +32,7 @@ export const StateToJsonSettings = (state: RootState): Settings => {
     'language',
     'dateFormat',
     'virtualDesktopStrategy',
-    'betaChannel',
+    'updater',
     'wall',
     'launcher',
   ]);