+++ src/background/error_handler.rs
@@ -49,6 +49,7 @@ define_app_errors!(
     Base64Decode(base64::DecodeError);
     WideStringNull(widestring::error::MissingNulTerminator);
     Reqwest(tauri_plugin_http::reqwest::Error);
+    Updater(tauri_plugin_updater::Error);
     WinScreenshot(win_screenshot::capture::WSError);
     EvalExpr(evalexpr::EvalexprError);
 );