+++ src/background/utils/mod.rs
@@ -1,6 +1,7 @@
 pub mod ahk;
 pub mod constants;
 pub mod pwsh;
+pub mod updater;
 pub mod virtual_desktop;
 mod winver;